llm-extract-image/requirements.txt
streamlit 
openai
pillow 
rembg 
numpy
opencv-python
onnxruntime
requests
matplotlib
bcrypt
fastapi
uvicorn
python-multipart
